+++ dflib-echarts/src/main/java/org/dflib/echarts/SeriesModelMaker.java
@@ -46,7 +46,7 @@ class SeriesModelMaker {
                 null,
                 null,
                 null,
-                null
+                sb.seriesOpts.itemStyle != null ? sb.seriesOpts.itemStyle.resolve() : null
         );
     }